Add files via upload
After buying all the components required for the project , the testing is done. The soil moisture sensor is tested by using Arduino and the results were monitored successfully. The motor driver L298N and the  Temperature and humidity sensor DHT22 are tested by using Arduino and corresponding results were obtained. The LDR sensor can be tested by using a multimeter and the readings were recorded.