Add files via upload
Individual presentation for the day is to design the schematic diagram of our project design. Here, I have used the fritzing software, which is an open source software to do the schematic design. By using fritzing, we can develop the schematic design, PCB layout as well as we have to the option to see  the breadboard view also.  The designs can be exported as image files too.